move django schemas to correct file
