+++ backend/ohq/schemas.py
@@ -36,3 +36,59 @@ class MassInviteSchema(AutoSchema):
             }
         }
         return operation
+
+
+class EventSchema(AutoSchema):
+    def get_operation(self, path, method):
+        op = super().get_operation(path, method)
+        if op["operationId"] == "listEvents":
+            op["parameters"].append(
+                {
+                    "name": "course",
+                    "in": "query",
+                    "required": True,
+                    "description": "A series of api/events/?course=1&course=2 "
+                    + "- where the numbers are the course pks",
+                    "schema": {"type": "string"},
+                }
+            )
+        return op
+
+
+class OccurrenceSchema(AutoSchema):
+    def get_operation(self, path, method):
+        op = super().get_operation(path, method)
+        if op["operationId"] == "listOccurrences":
+            op["parameters"].append(
+                {
+                    "name": "course",
+                    "in": "query",
+                    "required": True,
+                    "description": "A series of api/occurrences/?course=1&course=2 "
+                    + "- where the numbers are the course pks",
+                    "schema": {"type": "string"},
+                }
+            )
+            op["parameters"].append(
+                {
+                    "name": "filter_start",
+                    "in": "query",
+                    "required": True,
+                    "description": "The start date of the filter in ISO format in UTC+0.<br>"
+                    + "The returned events will have start_time strictly within "
+                    + "the range of the filter<br>"
+                    + "e.g 2021-10-05T12:41:37Z",
+                    "schema": {"type": "datetime"},
+                }
+            )
+            op["parameters"].append(
+                {
+                    "name": "filter_end",
+                    "in": "query",
+                    "required": True,
+                    "description": "The end date of the filter in ISO format in UTC+0<br>"
+                    + "e.g 2021-10-05T12:41:37Z",
+                    "schema": {"type": "datetime"},
+                }
+            )
+        return op
